requirements.txt
torch>=2.0
numpy
tqdm
pyyaml
matplotlib